library project

+++ clients/controllers/companies.js
@@ -5,10 +5,21 @@ const db = require('../db/connection');
 const validator = require('validator');
 
 app.get('/add-company', (req, res) => {
+
+    if(!req.session.auth) {
+        res.redirect('/');
+        return;
+    }
+
     res.render('template/company/add-company');
 });
 
 app.post('/add-company', (req, res) => {
+
+    if(!req.session.auth) {
+        res.redirect('/');
+        return;
+    }
     
     let companyName     = req.body.name;
     let companyAddress  = req.body.address;
@@ -49,6 +60,11 @@ app.post('/add-company', (req, res) => {
 
 app.get('/list-companies', (req, res) => {
 
+    if(!req.session.auth) {
+        res.redirect('/');
+        return;
+    }
+
     let messages = req.query.m;
     let status = req.query.s;
 
@@ -60,6 +76,12 @@ app.get('/list-companies', (req, res) => {
 });
 
 app.get('/edit-company/:id', (req, res) =>{
+
+    if(!req.session.auth) {
+        res.redirect('/');
+        return;
+    }
+
     let id = req.params.id;
     let messages = req.query.m;
     let status = req.query.s;
@@ -72,6 +94,12 @@ app.get('/edit-company/:id', (req, res) =>{
 });
 
 app.post('/edit-company/:id', (req, res) => {
+
+    if(!req.session.auth) {
+        res.redirect('/');
+        return;
+    }
+
     let id = req.params.id;
     let companyName     = req.body.name;
     let companyAddress  = req.body.address;
@@ -96,6 +124,12 @@ app.post('/edit-company/:id', (req, res) => {
 });
 
 app.get('/delete-company/:id', (req, res) => {
+
+    if(!req.session.auth) {
+        res.redirect('/');
+        return;
+    }
+    
     let id = req.params.id;
 
     db.query(`DELETE FROM companies WHERE id = '${id}'`, (err, resp) => {
